% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{extract_year_from_url}
\alias{extract_year_from_url}
\title{Extract Year from Bundestag Report URL}
\usage{
extract_year_from_url(x)
}
\arguments{
\item{x}{A character vector of URLs containing the year}
}
\value{
A character vector of extracted 4-digit years. If the URL does not
contain a year in the specified format, the result will be an empty string.
}
\description{
This function extracts the 4-digit year from a Bundestag party financing report URL.
}
